+++ lib/lastfm_archive/behaviour/archive.ex
@@ -30,9 +30,9 @@ defmodule LastfmArchive.Behaviour.Archive do
   @callback describe(user(), options()) :: {:ok, metadata()} | {:error, term()}
 
   @doc """
-  Optionally applies post-archive side effects such as archive transformation or loading.
+  Optionally applies post-archive side effects such as transformation into columnar storage.
   """
-  @callback after_archive(metadata(), transformer(), options()) :: {:ok, metadata()} | {:error, term()}
+  @callback post_archive(metadata(), transformer(), options()) :: {:ok, metadata()} | {:error, term()}
 
   @doc """
   Read access to the archive, returns an Explorer DataFrame for further data manipulation.
@@ -44,7 +44,7 @@ defmodule LastfmArchive.Behaviour.Archive do
   """
   @callback update_metadata(metadata(), options) :: {:ok, metadata()} | {:error, term()}
 
-  @optional_callbacks archive: 3, after_archive: 3
+  @optional_callbacks archive: 3, post_archive: 3
 
   defmacro __using__(_opts) do
     quote do