+++ test/lastfm_archive/archive/derived_archive_test.exs
@@ -4,12 +4,12 @@ defmodule LastfmArchive.Archive.DerivedArchiveTest do
   import ExUnit.CaptureLog
   import Fixtures.Archive
   import Hammox
+  import LastfmArchive.Archive.Transformers.TransformerSettings, only: [transformer: 1]
   import LastfmArchive.Utils, only: [user_dir: 1, metadata_filepath: 2]
 
   alias LastfmArchive.Archive.DerivedArchive
   alias LastfmArchive.Archive.FileArchiveMock
   alias LastfmArchive.Archive.Metadata
-  alias LastfmArchive.Archive.Transformers.FileArchiveTransformer
   alias LastfmArchive.FileIOMock
 
   alias Explorer.DataFrame
@@ -32,11 +32,12 @@ defmodule LastfmArchive.Archive.DerivedArchiveTest do
     %{file_archive_metadata: metadata, user: user}
   end
 
-  describe "after_archive/3 transform FileArchive" do
-    for format <- DerivedArchive.formats() do
-      test "into #{format} file", %{file_archive_metadata: metadata} do
+  describe "post_archive/3 transform" do
+    for format <- DerivedArchive.formats(), facet <- DerivedArchive.facets() do
+      test "#{facet} into #{format} file", %{file_archive_metadata: metadata} do
+        facet = unquote(facet)
         format = unquote(format)
-        metadata = metadata |> new_derived_archive_metadata(format: format)
+        metadata = metadata |> new_derived_archive_metadata(format: format, facet: facet)
 
         # 16 read for 16 months, each with 105 scrobbles
         FileArchiveMock
@@ -54,18 +55,20 @@ defmodule LastfmArchive.Archive.DerivedArchiveTest do
         end
 
         capture_log(fn ->
-          assert {:ok, _metadata} = DerivedArchive.after_archive(metadata, FileArchiveTransformer, format: format)
+          assert {:ok, _metadata} =
+                   DerivedArchive.post_archive(metadata, transformer(facet), format: format, facet: facet)
         end)
       end
     end
   end
 
   describe "describe/2" do
-    for format <- DerivedArchive.formats() do
-      test "existing #{format} derived archive", %{user: user, file_archive_metadata: metadata} do
+    for format <- DerivedArchive.formats(), facet <- DerivedArchive.facets() do
+      test "existing #{format} derived #{facet} archive", %{user: user, file_archive_metadata: metadata} do
+        facet = unquote(facet)
         format = unquote(format)
 
-        derived_archive_metadata = metadata |> new_derived_archive_metadata(format: format)
+        derived_archive_metadata = metadata |> new_derived_archive_metadata(format: format, facet: facet)
         derived_archive_metadata_filepath = metadata_filepath(user, format: format)
         file_archive_metadata_filepath = metadata_filepath(user, [])
         mimetype = DerivedArchive.mimetype(format)
@@ -84,7 +87,7 @@ defmodule LastfmArchive.Archive.DerivedArchiveTest do
                    identifier: ^user,
                    source: "local file archive",
                    title: "Lastfm archive of a_lastfm_user",
-                   type: DerivedArchive,
+                   type: ^facet,
                    extent: 400,
                    date: %{__struct__: Date},
                    temporal: {1_641_063_007, 1_680_547_807},
@@ -92,17 +95,18 @@ defmodule LastfmArchive.Archive.DerivedArchiveTest do
                  }
                } = DerivedArchive.describe(user, format: format)
 
-        assert description == "Lastfm archive of a_lastfm_user in #{format} format"
+        assert description == "Lastfm #{facet} archive of a_lastfm_user in #{format} format"
       end
 
-      test "#{format} returns new metadata when file archive exists", %{
+      test "returns metadata when archive exists for #{facet} archive in #{format}", %{
         user: user,
         file_archive_metadata: metadata
       } do
+        facet = unquote(facet)
         format = unquote(format)
 
         file_archive_metadata_filepath = metadata_filepath(user, [])
-        derived_archive_metadata_filepath = metadata_filepath(user, format: format)
+        derived_archive_metadata_filepath = metadata_filepath(user, format: format, facet: facet)
         mimetype = DerivedArchive.mimetype(format)
 
         FileIOMock
@@ -119,55 +123,67 @@ defmodule LastfmArchive.Archive.DerivedArchiveTest do
                    identifier: ^user,
                    source: "local file archive",
                    title: "Lastfm archive of a_lastfm_user",
-                   type: DerivedArchive,
+                   type: ^facet,
                    date: %{__struct__: Date},
                    extent: 400,
                    modified: _now,
                    temporal: {1_641_063_007, 1_680_547_807}
                  }
-               } = DerivedArchive.describe(user, format: format)
+               } = DerivedArchive.describe(user, format: format, facet: facet)
 
-        assert description == "Lastfm archive of a_lastfm_user in #{format} format"
+        assert description == "Lastfm #{facet} archive of a_lastfm_user in #{format} format"
       end
     end
   end
 
   describe "read/2" do
-    for format <- DerivedArchive.formats() do
-      test "#{format} returns data frame given a year option", %{user: user, file_archive_metadata: metadata} do
+    for format <- DerivedArchive.formats(), facet <- DerivedArchive.facets() do
+      test "data frame with a year option for #{facet} archive in #{format}", %{
+        user: user,
+        file_archive_metadata: metadata
+      } do
+        facet = unquote(facet)
         format = unquote(format)
-        metadata = metadata |> new_derived_archive_metadata(format: format)
-        opts = DerivedArchive.read_opts(format)
+        opts = [format: format, facet: facet]
 
-        filepath = Path.join([user_dir(user), "#{format}", "2023.#{format}"])
+        metadata = metadata |> new_derived_archive_metadata(format: format, facet: facet)
+        read_opts = DerivedArchive.read_opts(format)
+        archive_dir = "#{DerivedArchive.derived_archive_dir(opts)}"
+
+        filepath = Path.join([user_dir(user), archive_dir, "2023.#{format}"])
         filepath = if format == :csv, do: filepath <> ".gz", else: filepath
 
-        DataFrameMock |> expect(:"from_#{format}!", fn ^filepath, ^opts -> data_frame() end)
+        DataFrameMock |> expect(:"from_#{format}!", fn ^filepath, ^read_opts -> data_frame() end)
 
-        assert {:ok, %DataFrame{}} = DerivedArchive.read(metadata, year: 2023)
+        assert {:ok, %DataFrame{}} = DerivedArchive.read(metadata, year: 2023, facet: facet)
       end
 
-      test "#{format} when columns options is given", %{user: user, file_archive_metadata: metadata} do
+      test "#{facet} archive in #{format} with columns options", %{user: user, file_archive_metadata: metadata} do
+        facet = unquote(facet)
         format = unquote(format)
         columns = [:id, :album, :artist]
-        metadata = metadata |> new_derived_archive_metadata(format: format)
-        opts = DerivedArchive.read_opts(format) |> Keyword.put(:columns, columns)
+        opts = [format: format, facet: facet]
 
-        filepath = Path.join([user_dir(user), "#{format}", "2023.#{format}"])
+        metadata = metadata |> new_derived_archive_metadata(format: format, facet: facet)
+        read_opts = DerivedArchive.read_opts(format) |> Keyword.put(:columns, columns)
+        archive_dir = "#{DerivedArchive.derived_archive_dir(opts)}"
+
+        filepath = Path.join([user_dir(user), archive_dir, "2023.#{format}"])
         filepath = if format == :csv, do: filepath <> ".gz", else: filepath
 
-        DataFrameMock |> expect(:"from_#{format}!", fn ^filepath, ^opts -> data_frame() end)
+        DataFrameMock |> expect(:"from_#{format}!", fn ^filepath, ^read_opts -> data_frame() end)
 
-        assert {:ok, %DataFrame{}} = DerivedArchive.read(metadata, year: 2023, columns: columns)
+        assert {:ok, %DataFrame{}} = DerivedArchive.read(metadata, year: 2023, columns: columns, facet: facet)
       end
 
-      test "all #{format} (years) when no year option specified", %{file_archive_metadata: metadata} do
+      test "entire #{facet} archive in #{format} without year option", %{file_archive_metadata: metadata} do
+        facet = unquote(facet)
         format = unquote(format)
-        metadata = metadata |> new_derived_archive_metadata(format: format)
+        metadata = metadata |> new_derived_archive_metadata(format: format, facet: facet)
 
         # read all (2) years from files
         DataFrameMock |> expect(:"from_#{format}!", 2, fn _filepath, _opts -> data_frame() end)
-        assert {:ok, %DataFrame{}} = DerivedArchive.read(metadata, [])
+        assert {:ok, %DataFrame{}} = DerivedArchive.read(metadata, format: format, facet: facet)
       end
     end
   end