+++ test/lastfm_archive/behaviour/archive_test.exs
@@ -8,8 +8,10 @@ end
 # tests for default functions in the Archive behaviour
 defmodule LastfmArchive.Behaviour.ArchiveTest do
   use ExUnit.Case, async: true
+
   import Fixtures.Archive
   import Hammox
+  import LastfmArchive.Utils, only: [metadata_filepath: 2]
 
   alias LastfmArchive.Archive.Metadata
 
@@ -19,17 +21,19 @@ defmodule LastfmArchive.Behaviour.ArchiveTest do
     %{
       archive: LastfmArchive.TestArchive,
       metadata: file_archive_metadata("a_user"),
-      type: LastfmArchive.Behaviour.Archive.impl()
+      metadata_path: "test_data_dir/a_user/.metadata/scrobbles/json_archive",
+      type: :scrobbles
     }
   end
 
   describe "update_metadata/2" do
-    test "writes metadata to file", %{archive: archive, metadata: metadata, type: type} do
+    test "writes metadata to file", %{archive: archive, metadata: metadata, metadata_path: path, type: type} do
       metadata_encoded = Jason.encode!(metadata)
+      dir = path |> Path.dirname()
 
       LastfmArchive.FileIOMock
-      |> expect(:mkdir_p, fn "test_data_dir/a_user" -> :ok end)
-      |> expect(:write, fn "test_data_dir/a_user/.file_archive_metadata", ^metadata_encoded -> :ok end)
+      |> expect(:mkdir_p, fn ^dir -> :ok end)
+      |> expect(:write, fn ^path, ^metadata_encoded -> :ok end)
 
       assert {
                :ok,
@@ -46,13 +50,13 @@ defmodule LastfmArchive.Behaviour.ArchiveTest do
              } = archive.update_metadata(metadata, data_dir: "test_data_dir")
     end
 
-    test "reset an existing archive via 'overwrite' option", %{archive: archive} do
+    test "reset an existing archive via 'reset' option", %{archive: archive} do
       earlier_created_datetime = DateTime.add(DateTime.utc_now(), -3600, :second)
       metadata = file_archive_metadata("a_user", earlier_created_datetime)
 
       LastfmArchive.FileIOMock
-      |> expect(:mkdir_p, fn "existing_archive/a_user" -> :ok end)
-      |> expect(:write, fn "existing_archive/a_user/.file_archive_metadata", _ -> :ok end)
+      |> expect(:mkdir_p, fn _dir -> :ok end)
+      |> expect(:write, fn _path, _ -> :ok end)
 
       assert {
                :ok,
@@ -74,11 +78,13 @@ defmodule LastfmArchive.Behaviour.ArchiveTest do
   end
 
   describe "describe/2" do
-    test "an existing file archive", %{archive: archive, metadata: metadata, type: type} do
-      archive_id = metadata.creator
-      metadata_path = Path.join([Application.get_env(:lastfm_archive, :data_dir), archive_id, ".file_archive_metadata"])
+    setup context do
+      %{metadata_path: metadata_filepath(context.metadata.creator, [])}
+    end
 
-      LastfmArchive.FileIOMock |> expect(:read, fn ^metadata_path -> {:ok, metadata |> Jason.encode!()} end)
+    test "an existing file archive", %{archive: archive, metadata: metadata, metadata_path: path, type: type} do
+      user = metadata.creator
+      LastfmArchive.FileIOMock |> expect(:read, fn ^path -> {:ok, metadata |> Jason.encode!()} end)
 
       assert {
                :ok,
@@ -96,7 +102,7 @@ defmodule LastfmArchive.Behaviour.ArchiveTest do
                  temporal: {1_617_303_007, 1_617_475_807},
                  modified: "2023-06-09T14:36:16.952540Z"
                }
-             } = archive.describe(archive_id)
+             } = archive.describe(user)
     end
 
     test "returns new metadata for non-existing archive", %{archive: archive, type: type} do